feat: move calcTotalPrice function to utils module and update cart service
